pipeline/requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pypdf>=4.0.0
